backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
httpx==0.27.2
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
google-cloud-aiplatform==1.68.0
elasticsearch==8.15.1
aiohttp==3.10.10
prometheus-client==0.21.0
